moved validation function to separate file
